+++ api/jest.config.js
@@ -1,10 +1,8 @@
 const config = {
   // verbose: true,
   roots: ["src/"],
+  testEnvironment: "node",
   testPathIgnorePatterns: ["/node_modules/", "/__mocks__/", "/helpers/", "/fixtures/"],
-  transform: {
-    "^.+\\.jsx?$": "babel-jest",
-  },
 };
 
 module.exports = config;
