+++ api/package.json
@@ -11,7 +11,7 @@
     "dev": "nodemon ./src/index.js",
     "start": "npm run install:gm && node ./src/index.js",
     "install:gm": "cave resolve media-gfx/GraphicsMagick",
-    "test": "jest --silent",
+    "test": "jest --silent --runInBand",
     "coverage": "jest --runInBand --coverage --silent"
   },
   "author": "SELEGO",
